[Refactor] 컴포넌트 분리
- 단일 파일에 너무 많은 컴포넌트가 작성되어 있어 컴포넌트 분리
- 코드 가독성이 여전히 떨어지는 상황이라 추가 리팩토링 예정

Issues #122
